Support for upcoming smart-contract-based allocators.
What: this PR stores "address" and "tooling" fields from allocator JSON
files in database and exposes to clients via API.

Why: smart-contract allocators will need to make a different call
on-chain to grant datacap to a client. Instead of calling verifreg
directly, they'll need to call smart contract which will do it on their
behalf. Address of this smart contract is stored in the "address" field
and we need the "tooling" field to know when to use the new approach.